Update method signatures in SlugService
The function slugify now accepts a string parameter, and the updateSlug method accepts an entityId parameter. Also, a check has been added in the updateSlug method to ensure that the entityId, if present, does not match the existing id before updating the slug. A counter was also initialised to 1 instead of 2 for generating new slugs.