Add password reset verification template
Implemented a password reset verification template for the blog project. The form informs the user that a reset link may have been sent to their email if a corresponding account was located. It notifies users that the link will expire in 10 minutes and mentions that they should check their spam folder if an email is not received.