Update routing for expanded blog and comment administration
The routing in 'index.php' has been updated to include new routes for blog creation, editing, and deletion, as well as comment validation and deletion. This will make the administration of blogs and comments more efficient and user-friendly. These changes are integral to improving overall blog post utility and user navigation in the blog overview page.