Refactor MailerService class to use Symfony Mailer
The MailerService class was heavily refactored to begin using the Symfony Mailer component instead of Swift_Mailer. This includes a change in the way a mail transport is created, the way mails are constructed, and exception handling. It also includes validation for reply email and an exception for an invalid email address.