Handle mail exceptions in ResetPasswordController
The ResetPasswordController is now set to throw a TransportExceptionInterface in case sending email fails. This change comes in line with the migration from Swift_Mailer to Symfony Mailer in the MailerService class. The way email transport is created, and mails are constructed was also updated, along with the UserRepository call method.