Replace Swift_DependencyException with TransportExceptionInterface
The Swift_DependencyException has been replaced with Symfony's TransportExceptionInterface in the RegisterController to handle potential email sending failures. Swift DependencyExceptions are removed and replaced to accord with the switch from Swift_Mailer to Symfony Mailer in our MailerService implementation. The construction of emails and UserRepository method call were updated accordingly.