Update UI in reset password pages
This commit refines the user interface in the reset_password.html.twig and check_email.html.twig pages. Changes include adding margins, changing the border style of the card, and replacing the placeholder image with the actual logo. The modifications overall enhance the visual appeal and usability of the pages.