Update styles in HTML templates
This commit makes minor modifications in the home, about, and blog HTML templates to improve visual presentation. Changes include removing a punctuation mark from the blog template, adding a "bg-white" class to the about and blog sections for a cleaner look, and adding a "shadow" class to the blog cards for a subtle depth effect.