Refactor conditional statements in CommentController
The code has been updated to use double ampersand (&&) instead of 'and' for logical AND operations. This change has been applied in two places in the CommentController, specifically in the condition to check user permissions for editing and deleting comments. This is part of a larger effort to adhere to best coding practices.